Split cli from client and set up subpackages
