+++ client/requirements.txt
@@ -1,0 +1,3 @@
+cryptography
+requests
+structlog
