+++ client/template.conf
@@ -1,0 +1,4 @@
+[teos]
+rpc_connect = localhost
+rpc_port = 9000
+
